requirements.txt
openai
python-dotenv
SpeechRecognition
pyttsx3
numpy
gtts
pyaudio